Consolidate create_grid_from_shape into geo/grid.py
The two copies were byte-identical, MIN_DISTANCE_METRES included, and neither had an in-library caller -- only tests and notebooks. Their tests merge into tests/geo/test_grid.py.
