+++ climate_risk/geo/grid.py
@@ -1,0 +1,73 @@
+import geopandas as gpd
+import numpy as np
+import pandas as pd
+
+from climate_risk.geo.crs import GEOGRAPHIC_CRS
+from climate_risk.geo.distance import get_distance_to
+
+# get_distance_to reprojects to PROJECTED_CRS, so distances are in metres.
+MIN_DISTANCE_METRES = 1.0
+
+
+def create_grid_from_shape(
+    shapefile: gpd.GeoDataFrame,
+    rivers: gpd.GeoDataFrame,
+    coastline: gpd.GeoDataFrame,
+    grid_size: int = 100,
+) -> gpd.GeoDataFrame:
+    """
+    Lay a regular grid over the bounds of ``shapefile`` and measure each point to river and coast.
+
+    Parameters
+    ----------
+    shapefile : GeoDataFrame
+        Region to cover. Points falling outside its geometry are dropped, and an ``ISO_A3`` column,
+        if present, labels the survivors.
+    rivers : GeoDataFrame
+        River geometries, carrying ``ORD_FLOW`` and ``HYRIV_ID``.
+    coastline : GeoDataFrame
+        Coastline geometries. Distance is measured to the boundary.
+    grid_size : int, optional
+        Points per axis, so the grid holds ``grid_size ** 2`` before clipping. Default 100.
+
+    Returns
+    -------
+    GeoDataFrame
+        One row per surviving point, with ``long``, ``lat``, distances to the nearest river and
+        coastline, their logs, and ``ISO``.
+    """
+    long_min, lat_min, long_max, lat_max = shapefile.dissolve().bounds.values.ravel()
+    long_grid = np.linspace(long_min, long_max, grid_size)
+    lat_grid = np.linspace(lat_min, lat_max, grid_size)
+
+    grid = np.column_stack([x.ravel() for x in np.meshgrid(long_grid, lat_grid)])
+    grid = gpd.GeoSeries(gpd.points_from_xy(*grid.T), crs=GEOGRAPHIC_CRS)
+    grid = gpd.GeoDataFrame({"geometry": grid})
+
+    point_overlay = grid.overlay(shapefile, how="intersection")
+    points = point_overlay.geometry
+    points = points.to_frame().assign(long=lambda x: x.geometry.x, lat=lambda x: x.geometry.y)
+
+    distances_to_rivers = get_distance_to(rivers, points=points, return_columns=["ORD_FLOW", "HYRIV_ID"]).rename(
+        columns={"distance_to_closest": "distance_to_river"}
+    )
+
+    points = pd.merge(points, distances_to_rivers, left_index=True, right_index=True, how="left")
+
+    distances_to_coastlines = get_distance_to(coastline.boundary, points=points, return_columns=None).rename(
+        columns={"distance_to_closest": "distance_to_coastline"}
+    )
+
+    points = pd.merge(points, distances_to_coastlines, left_index=True, right_index=True, how="left")
+
+    points = points.assign(
+        log_distance_to_river=lambda x: np.log(x.distance_to_river.clip(lower=MIN_DISTANCE_METRES)),
+        log_distance_to_coastline=lambda x: np.log(x.distance_to_coastline.clip(lower=MIN_DISTANCE_METRES)),
+    )
+
+    if "ISO_A3" in point_overlay.columns:
+        points["ISO"] = point_overlay.ISO_A3
+    else:
+        points["ISO"] = "LAO"
+
+    return points
